Add CreateProjectPort
Also rename the packages for the ports "driven" and "driver".

Driver ports are the API of the application (used by driver adapters).
Driven ports are service provider interfaces and are used for business logic (implemented by driven Adapters).
