+++ coderadar-server/coderadar-core/src/main/java/org/wickedsource/coderadar/project/rest/CreateProjectController.java
@@ -14,8 +14,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.wickedsource.coderadar.projectadministration.port.in.CreateProjectCommand;
-import org.wickedsource.coderadar.projectadministration.port.in.CreateProjectUseCase;
+import org.wickedsource.coderadar.projectadministration.port.driver.CreateProjectCommand;
+import org.wickedsource.coderadar.projectadministration.port.driver.CreateProjectUseCase;
 
 @Controller
 @Transactional
