+++ coderadar-server/coderadar-core/src/main/java/org/wickedsource/coderadar/projectadministration/port/driven/CreateProjectPort.java
@@ -1,0 +1,7 @@
+package org.wickedsource.coderadar.projectadministration.port.driven;
+
+import org.wickedsource.coderadar.project.domain.Project;
+
+public interface CreateProjectPort {
+  Project createProject(Project project);
+}
