+++ coderadar-server/coderadar-core/src/main/java/org/wickedsource/coderadar/projectadministration/port/driver/CreateProjectUseCase.java
@@ -1,4 +1,4 @@
-package org.wickedsource.coderadar.projectadministration.port.in;
+package org.wickedsource.coderadar.projectadministration.port.driver;
 
 public interface CreateProjectUseCase {
   CreateProjectCommand createProject(CreateProjectCommand command);
